+++ demo/src/index.js
@@ -1,5 +1,6 @@
 import remarklet from "@zw/remarklet";
-// remarklet.options({
-//     persist: true,
-// });
+remarklet.options({
+    persist: true,
+    hide: true,
+});
 remarklet.activate();