+++ CHANGELOG.md
@@ -6,7 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ## [Unreleased]
 
-No unreleased changes.
+### Changed
+
+- Added a way for users to hide elements from view.
 
 ## [1.1.1] - 2025-04-13